requirements.txt
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
folium>=0.14.0
requests>=2.31.0
openai>=1.0.0 